Printed spring y carriage is backwards compatible
The PTFE bushing carriage ended up having a bunch of chatter, and was
kind of hard to put together. This seems promising.

I used some UHMW PE tape on the contact surfaces to avoid the printed
surfaces (PLA) wearing out.

We'll see how the PLA reacts to being under constant compression and if
it manages to hold its shape enough.