A *much* simpler plotter design
Driven only on one side, with one axis carrying the motor for the other
axes (Y is carried by X, in this case)

Committing before trying a block-and-tackle model for the X axis to get
the X motor off of the carriage and possibly give more strength to the X
motor.

Hopefully I can resist the block-and-tackle, though, because:
A) YAGNI
B) more length of cheap GT2 belt = more stretching?