A different take on a Y axis
Maybe this is more balanced, with weight on both sides of the xy joiner?

At the very least, the belt seems more in line with center of mass of
the YZ carriage, and is closer to the aluminum extrusion, too.

Need to add belt teeth to the YZ carriage, but that _should_ be
straightforward?

Still need to figure out how to attach the X belt. I've been
procrastinating on that one. :(

I like how it's looking, though, and I could probably add a hook to the
Y motor area where I could attach some sort of cable chain